requirements.txt
Django
whitenoise
pandas
yfinance
plotly
numpy
pytz
pandas_market_calendars
alpaca-py
Celery
pandas_ta 
lightweight-charts
django-celery-beat
eventlet # Needed for Celery on Windows
azure-servicebus
azure-identity
